Make ContentBuilder generic and remove the other builders
