+++ Sources/HTMLKit/External/Elements/TableElements.swift
@@ -43,158 +43,19 @@ public typealias Th = HeaderCell
 /// The element
 ///
 ///
-public struct Table: ContentNode, TableElement {
-
-    internal var name: String { "table" }
-
-    internal var attributes: OrderedDictionary<String, Any>?
-
-    internal var content: AnyContent
-
-    public init(@ContentBuilder content: () -> AnyContent) {
-        self.content = content()
-    }
-    
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
-        self.attributes = attributes
-        self.content = content
-    }
-}
-
-extension Table: GlobalAttributes, WidthAttribute, HeightAttribute {
-    
-    public func accessKey(_ value: String) -> Table {
-        return mutate(accesskey: value)
-    }
-
-    public func autocapitalize(_ type: Capitalization) -> Table {
-        return mutate(autocapitalize: type.rawValue)
-    }
-
-    public func autofocus() -> Table {
-        return mutate(autofocus: "autofocus")
-    }
-
-    public func `class`(_ value: String) -> Table {
-        return mutate(class: value)
-    }
-
-    public func isEditable(_ condition: Bool) -> Table {
-        return mutate(contenteditable: condition)
-    }
-
-    public func direction(_ type: Direction) -> Table {
-        return mutate(dir: type.rawValue)
-    }
-
-    public func isDraggable(_ condition: Bool) -> Table {
-        return mutate(draggable: condition)
-    }
-
-    public func enterKeyHint(_ type: Hint) -> Table {
-        return mutate(enterkeyhint: type.rawValue)
-    }
-
-    public func hidden() -> Table {
-        return mutate(hidden: "hidden")
-    }
-
-    public func inputMode(_ value: String) -> Table {
-        return mutate(inputmode: value)
-    }
-
-    public func `is`(_ value: String) -> Table {
-        return mutate(is: value)
-    }
-
-    public func itemId(_ value: String) -> Table {
-        return mutate(itemid: value)
-    }
-
-    public func itemProperty(_ value: String) -> Table {
-        return mutate(itemprop: value)
-    }
-
-    public func itemReference(_ value: String) -> Table {
-        return mutate(itemref: value)
-    }
-
-    public func itemScope(_ value: String) -> Table {
-        return mutate(itemscope: value)
-    }
-
-    public func id(_ value: String) -> Table {
-        return mutate(id: value)
-    }
-
-    public func language(_ type: Language) -> Table {
-        return mutate(lang: type.rawValue)
-    }
-
-    public func nonce(_ value: String) -> Table {
-        return mutate(nonce: value)
-    }
-
-    public func role(_ value: String) -> Table {
-        return mutate(role: value)
-    }
-
-    public func hasSpellCheck(_ condition: Bool) -> Table {
-        return mutate(spellcheck: condition)
-    }
-
-    public func style(_ value: String) -> Table {
-        return mutate(style: value)
-    }
-
-    public func tabIndex(_ value: String) -> Table {
-        return mutate(tabindex: value)
-    }
-
-    public func title(_ value: String) -> Table {
-        return mutate(title: value)
-    }
-
-    public func translate(_ value: String) -> Table {
-        return mutate(translate: value)
-    }
-
-    public func width(_ size: Int) -> Table {
-        return mutate(width: size)
-    }
-    
-    public func height(_ size: Int) -> Table {
-        return mutate(height: size)
-    }
-}
-
-extension Table: AnyContent {
-    
-    public func prerender(_ formula: Renderer.Formula) throws {
-        try self.build(formula)
-    }
-    
-    public func render<T>(with manager: Renderer.ContextManager<T>) throws -> String {
-        try self.build(with: manager)
-    }
-}
-
-/// The element
-///
-///
 public struct Caption: ContentNode, TableElement {
 
     internal var name: String { "caption" }
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -320,13 +181,13 @@ public struct ColumnGroup: ContentNode, TableElement {
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -455,13 +316,13 @@ public struct Column: ContentNode, TableElement {
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -590,13 +451,13 @@ public struct TableBody: ContentNode, TableElement {
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -729,13 +590,13 @@ public struct TableHead: ContentNode, TableElement {
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -868,13 +729,13 @@ public struct TableFoot: ContentNode, TableElement {
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -999,13 +860,13 @@ public struct TableRow: ContentNode, TableElement {
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -1138,13 +999,13 @@ public struct DataCell: ContentNode, TableElement {
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -1281,13 +1142,13 @@ public struct HeaderCell: ContentNode, TableElement {
 
     internal var attributes: OrderedDictionary<String, Any>?
 
-    internal var content: AnyContent
+    internal  var content: [AnyContent]
 
-    public init(@ContentBuilder content: () -> AnyContent) {
+    public init(@ContentBuilder<AnyContent> content: () -> [AnyContent]) {
         self.content = content()
     }
     
-    internal init(attributes: OrderedDictionary<String, Any>?, content: AnyContent) {
+    internal init(attributes: OrderedDictionary<String, Any>?, content: [AnyContent]) {
         self.attributes = attributes
         self.content = content
     }
@@ -1405,17 +1266,6 @@ extension HeaderCell: GlobalAttributes, ColumnSpanAttribute, RowSpanAttribute, H
     
     public func scope(_ value: String) -> HeaderCell {
         return mutate(scope: value)
-    }
-}
-
-extension HeaderCell: Localizable {
-    
-    public init(_ localizedKey: String) {
-        self.content = Localized(key: localizedKey)
-    }
-
-    public init<B>(_ localizedKey: String, with context: TemplateValue<B>) where B : Encodable {
-        self.content = Localized(key: localizedKey, context: context)
     }
 }
