--- Sources/HTMLKit/Internal/Core/Builders/StringBuilder.swift
@@ -1,9 +1,0 @@
-/// The builder is for
-///
-///
-@resultBuilder public struct StringBuilder {
-    
-    public static func buildBlock(_ components: String...) -> String {
-        return components.joined()
-    }
-}
