+++ Sources/HTMLKit/Internal/Core/Nodes.swift
@@ -11,9 +11,9 @@ internal protocol ContentNode: AnyNode {
     
     var attributes: OrderedDictionary<String, Any>? { get }
     
-    var content: Content { get }
+    var content: [Content] { get }
     
-    init(attributes: OrderedDictionary<String, Any>?, content: Content)
+    init(attributes: OrderedDictionary<String, Any>?, content: [Content])
 }
 
 extension ContentNode {
@@ -27,6 +27,426 @@ extension ContentNode {
 }
 
 extension ContentNode where Content == AnyContent {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == BodyElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == DescriptionElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == FigureElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == FormElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == BasicElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == HeadElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == InputElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == ListElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == MapElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == MediaElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == ObjectElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == RubyElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == TableElement {
+    
+    internal func build(_ formula: Renderer.Formula) throws {
+
+        formula.add(string: "<\(name)")
+
+        if let attributes = attributes {
+
+            attributes.forEach { attribute in
+                formula.add(string: " \(attribute.key)=\"\(attribute.value)\"")
+            }
+        }
+
+        formula.add(string: ">")
+
+        try content.prerender(formula)
+
+        formula.add(string: "</\(name)>")
+    }
+
+    internal func build<T>(with manager: Renderer.ContextManager<T>) throws -> String {
+
+        guard let attributes = attributes else {
+            return try "<\(name)>\(content.render(with: manager))</\(name)>"
+        }
+        
+        return try "<\(name)" + attributes.map { attribute in return " \(attribute.key)=\"\(attribute.value)\"" } + ">\(content.render(with: manager))</\(name)>" as! String
+    }
+}
+
+extension ContentNode where Content == HtmlElement {
     
     internal func build(_ formula: Renderer.Formula) throws {
 
@@ -71,7 +491,7 @@ extension ContentNode where Content == String {
         
         formula.add(string: ">")
         
-        formula.add(string: content)
+        try content.prerender(formula)
         
         formula.add(string: "</\(name)>")
     }
