+++ Tests/HTMLKitTests/ComponentTests.swift
@@ -5,7 +5,7 @@ final class ComponentTests: XCTestCase {
     
     struct TestPage: Page {
         
-        @ContentBuilder var body: AnyContent
+        @ContentBuilder<AnyContent> var body: AnyContent
     }
     
     var renderer = Renderer()
